Admin Gui change characteristics to view works
